Ephemeral defer + Followup returns an ephemeral followup
