+++ core/src/main/kotlin/event/interaction/InteractionCreate.kt
@@ -2,9 +2,8 @@ package dev.kord.core.event.interaction
 
 import dev.kord.common.annotation.KordPreview
 import dev.kord.core.Kord
-import dev.kord.core.entity.interaction.Interaction
 import dev.kord.core.event.Event
-
+import dev.kord.core.entity.interaction.Interaction
 @KordPreview
 class InteractionCreateEvent(
     val interaction: Interaction,
